pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
authors = ["Pavel Ilchenko <arisesinmight@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]


[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flake8 = "^7.1.1"
pytest = "^8.3.4"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
validators = "^0.34.0"
psycopg2-binary = "^2.9.10"
isort = "^5.13.2"
datetime = "^5.5"
psycopg-pool = "^3.2.4"
psycopg = {extras = ["pool"], version = "^3.2.3"}
requests = "^2.32.3"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
